requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
pydantic-settings==2.10.1
python-dotenv==1.0.0
sqlalchemy==2.0.42
email-validator==2.1.1
passlib==1.7.4
bcrypt==3.2.2